classification_main/requirements.txt
Flask==1.1.1
fuzzywuzzy==0.17.0
# ipython==7.8.0
# matplotlib==3.1.1
numpy==1.17.1
# openpyxl==3.0.0
# packaging==19.2
pandas==0.25.1
# PyYAML==5.3.1
# six==1.12.0
# urllib3==1.25.3
# Werkzeug==0.15.6
# xlrd==1.2.0
requests==2.23.0
# pytest==5.4.1
json-logging==1.2.0
gunicorn==20.0.4
flask-nameko==1.4.0
nameko==2.12.0
Cython==0.29.21
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
iteration-utilities==0.11.0
neo4j==4.3.1
nltk==3.5


